requirements.txt
alembic==1.16.5
annotated-types==0.7.0
bcrypt==4.3.0
greenlet==3.2.4
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
SQLAlchemy==2.0.43
typing-inspection==0.4.1
typing_extensions==4.15.0
websockets==15.0.1
